requirements.txt
numpy
pandas
scikit-learn
scipy
torch
pillow
